# openspec/changes/sdlc-visualizer/interface-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Arsitect SDLC Visualizer API
  version: 1.0.0
  description: Arsitect 可视化驾驶舱 REST API 契约（基于详细设计 20/21 模块导出）
servers:
- url: /api/v1
  description: API v1 基线
paths:
...
  /templates/{level}:
    get:
      tags:
      - templates
      summary: Get Template
      description: Get a template with its stage sequence.
      operationId: get_template_api_v1_templates__level__get
      parameters:
      - name: level
        in: path
        required: true
        schema:
          type: string
          title: Level
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateDetailDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /templates/projects/{project_id}/stage-sequence:
    get:
      tags:
      - templates
      summary: Get Stage Sequence
      description: Get the stage sequence for a project.
      operationId: get_stage_sequence_api_v1_templates_projects__project_id__stage_sequence_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectStageDTO'
                title: Response Get Stage Sequence Api V1 Templates Projects  Project Id  Stage Sequence Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /templates/projects/{project_id}/template-deviation/preview:
    post:
      tags:
      - templates
      summary: Preview Template Deviation
      description: Preview the impact of switching to a different template.
      operationId: preview_template_deviation_api_v1_templates_projects__project_id__template_deviation_preview_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateDeviationPreviewRequestDTO'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateDeviationPreviewDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /templates/projects/{project_id}/template-deviation:
    post:
      tags:
      - templates
      summary: Confirm Template Deviation
      description: Confirm and apply template switch.
      operationId: confirm_template_deviation_api_v1_templates_projects__project_id__template_deviation_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateDeviationConfirmDTO'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Confirm Template Deviation Api V1 Templates Projects  Project Id  Template Deviation Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /complexity/templates/{level}:
    get:
      tags:
      - complexity
      summary: Get Template Recommendation
      description: Get template recommendation for a complexity level.
      operationId: get_template_recommendation_api_v1_complexity_templates__level__get
      parameters:
      - name: level
        in: path
        required: true
        schema:
          type: string
          title: Level
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplexityTemplateDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ComplexityTemplateDTO:
      properties:
        level:
          type: string
          title: Level
        label:
          type: string
          title: Label
        recommended_template:
          type: string
          title: Recommended Template
        description:
          type: string
          title: Description
        stage_count:
          type: integer
          title: Stage Count
        estimated_skill_count:
          type: integer
          title: Estimated Skill Count
      type: object
      required:
      - level
      - label
      - recommended_template
      - description
      - stage_count
      - estimated_skill_count
      title: ComplexityTemplateDTO
      description: DTO for complexity template recommendation.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProjectStageDTO:
      properties:
        project_stage_id:
          type: string
          title: Project Stage Id
        project_id:
          type: string
          title: Project Id
        stage_id:
          type: string
          title: Stage Id
        order_index:
          type: integer
          title: Order Index
        status:
          type: string
          title: Status
        primary_skill_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Primary Skill Id
        skippable:
          type: boolean
          title: Skippable
          default: false
        is_frozen:
          type: boolean
          title: Is Frozen
          default: false
        merge_group_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Merge Group Id
        execution_status:
          type: string
          title: Execution Status
      type: object
      required:
      - project_stage_id
      - project_id
      - stage_id
      - order_index
      - status
      - execution_status
      title: ProjectStageDTO
      description: DTO for a project stage.
...
    TemplateDetailDTO:
      properties:
        template:
          $ref: '#/components/schemas/TemplateResponseDTO'
        stages:
          items:
            $ref: '#/components/schemas/TemplateStageDTO'
          type: array
          title: Stages
      type: object
      required:
      - template
      - stages
      title: TemplateDetailDTO
      description: Template with its stage sequence.
    TemplateDeviationConfirmDTO:
      properties:
        new_template_id:
          type: string
          title: New Template Id
          description: 目标模板 ID
      type: object
      required:
      - new_template_id
      title: TemplateDeviationConfirmDTO
      description: Request to confirm template deviation.
    TemplateDeviationPreviewDTO:
      properties:
        frozen_count:
          type: integer
          title: Frozen Count
        removed_count:
          type: integer
          title: Removed Count
        added_count:
          type: integer
          title: Added Count
        retained_count:
          type: integer
          title: Retained Count
      type: object
      required:
      - frozen_count
      - removed_count
      - added_count
      - retained_count
      title: TemplateDeviationPreviewDTO
      description: Preview of template switch impact.
    TemplateDeviationPreviewRequestDTO:
      properties:
        new_template_id:
          type: string
          title: New Template Id
          description: 目标模板 ID
      type: object
      required:
      - new_template_id
      title: TemplateDeviationPreviewRequestDTO
      description: Request to preview template deviation.
    TemplateResponseDTO:
      properties:
        template_id:
          type: string
          title: Template Id
        template_name:
          type: string
          title: Template Name
        description:
          type: string
          title: Description
        stage_count:
          type: integer
          title: Stage Count
        estimated_skill_count:
          type: integer
          title: Estimated Skill Count
        applicable_complexity:
          type: string
          title: Applicable Complexity
        config_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Config Json
      type: object
      required:
      - template_id
      - template_name
      - description
      - stage_count
      - estimated_skill_count
      - applicable_complexity
      title: TemplateResponseDTO
      description: DTO for a template.
    TemplateStageDTO:
      properties:
        stage_id:
          type: string
          title: Stage Id
        stage_name:
          type: string
          title: Stage Name
        order_index:
          type: integer
          title: Order Index
        template_id:
          type: string
          title: Template Id
        primary_skill_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Primary Skill Id
        auxiliary_skill_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Auxiliary Skill Ids
        gate_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Gate Id
        skippable:
          type: boolean
          title: Skippable
          default: false
        merge_group_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Merge Group Id
      type: object
      required:
      - stage_id
      - stage_name
      - order_index
      - template_id
      title: TemplateStageDTO
      description: DTO for a template stage.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: MVP 阶段为本地会话占位，P1 升级为 JWT
...